+++ pkg/sql/executor_ir_test.go
@@ -1,0 +1,184 @@
+// Copyright 2019 The SQLFlow Authors. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sql
+
+import (
+	"fmt"
+	"io/ioutil"
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+const (
+	testTrainSelectIris = testSelectIris + `
+TO TRAIN DNNClassifier
+WITH
+  model.n_classes = 3,
+  model.hidden_units = [10, 20]
+COLUMN sepal_length, sepal_width, petal_length, petal_width
+LABEL class
+INTO sqlflow_models.my_dnn_model;
+`
+	testPredictSelectIris = `
+SELECT *
+FROM iris.test
+TO PREDICT iris.predict.class
+USING sqlflow_models.my_dnn_model;
+`
+	testClusteringTrain = `SELECT sepal_length, sepal_width, petal_length, petal_width
+FROM iris.train
+TO TRAIN sqlflow_models.DeepEmbeddingClusterModel
+WITH
+  model.pretrain_dims = [10,10],
+  model.n_clusters = 3,
+  model.pretrain_lr = 0.001,
+  train.batch_size = 1
+COLUMN sepal_length, sepal_width, petal_length, petal_width
+INTO sqlflow_models.my_clustering_model;
+`
+	testClusteringPredict = `
+SELECT sepal_length, sepal_width, petal_length, petal_width
+FROM iris.test
+TO PREDICT iris.predict.class
+USING sqlflow_models.my_clustering_model;
+`
+	testXGBoostTrainSelectIris = ` 
+SELECT *
+FROM iris.train
+TO TRAIN xgboost.gbtree
+WITH
+    objective="multi:softprob",
+    train.num_boost_round = 30,
+    eta = 0.4,
+    num_class = 3
+COLUMN sepal_length, sepal_width, petal_length, petal_width
+LABEL class 
+INTO sqlflow_models.my_xgboost_model;
+`
+	testAnalyzeTreeModelSelectIris = `
+SELECT * FROM iris.train
+TO EXPLAIN sqlflow_models.my_xgboost_model
+USING TreeExplainer;
+`
+	testXGBoostPredictIris = ` 
+SELECT *
+FROM iris.test
+TO PREDICT iris.predict.class
+USING sqlflow_models.my_xgboost_model;
+`
+)
+
+func TestExecuteXGBoost(t *testing.T) {
+	a := assert.New(t)
+	modelDir := ""
+	a.NotPanics(func() {
+		stream := parseAndRunSQL(testXGBoostTrainSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testAnalyzeTreeModelSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testXGBoostPredictIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+	})
+}
+
+func TestExecuteXGBoostRegression(t *testing.T) {
+	a := assert.New(t)
+	modelDir := ""
+	a.NotPanics(func() {
+		stream := parseAndRunSQL(testXGBoostTrainSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testAnalyzeTreeModelSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testXGBoostPredictIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+	})
+}
+
+func TestExecutorTrainAndPredictDNN(t *testing.T) {
+	a := assert.New(t)
+	modelDir := ""
+	a.NotPanics(func() {
+		stream := parseAndRunSQL(testTrainSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testPredictSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+	})
+}
+
+func TestExecutorTrainAndPredictClusteringLocalFS(t *testing.T) {
+	a := assert.New(t)
+	modelDir, e := ioutil.TempDir("/tmp", "sqlflow_models")
+	a.Nil(e)
+	defer os.RemoveAll(modelDir)
+	a.NotPanics(func() {
+		stream := parseAndRunSQL(testClusteringTrain, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testClusteringPredict, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+	})
+}
+
+func TestExecutorTrainAndPredictDNNLocalFS(t *testing.T) {
+	a := assert.New(t)
+	modelDir, e := ioutil.TempDir("/tmp", "sqlflow_models")
+	a.Nil(e)
+	defer os.RemoveAll(modelDir)
+	a.NotPanics(func() {
+		stream := parseAndRunSQL(testTrainSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+		stream = parseAndRunSQL(testPredictSelectIris, modelDir, testDB)
+		a.True(goodStream(stream.ReadAll()))
+	})
+}
+
+func TestExecutorTrainAndPredictionDNNClassifierDENSE(t *testing.T) {
+	if getEnv("SQLFLOW_TEST_DB", "mysql") == "hive" {
+		t.Skip(fmt.Sprintf("%s: skip Hive test", getEnv("SQLFLOW_TEST_DB", "mysql")))
+	}
+	a := assert.New(t)
+	a.NotPanics(func() {
+		trainSQL := `SELECT * FROM iris.train_dense
+TO TRAIN DNNClassifier
+WITH
+model.n_classes = 3,
+model.hidden_units = [10, 20],
+train.epoch = 200,
+train.batch_size = 10,
+train.verbose = 1
+COLUMN NUMERIC(dense, 4)
+LABEL class
+INTO sqlflow_models.my_dense_dnn_model;`
+		stream := parseAndRunSQL(trainSQL, "", testDB)
+		a.True(goodStream(stream.ReadAll()))
+
+		predSQL := `SELECT * FROM iris.test_dense
+TO PREDICT iris.predict_dense.class
+USING sqlflow_models.my_dense_dnn_model
+;`
+		stream = parseAndRunSQL(predSQL, "", testDB)
+		a.True(goodStream(stream.ReadAll()))
+	})
+}
+
+func TestCreatePredictionTable(t *testing.T) {
+	a := assert.New(t)
+	trainParsed, e := newParser().Parse(testTrainSelectIris)
+	a.NoError(e)
+	predParsed, e := newParser().Parse(testPredictSelectIris)
+	a.NoError(e)
+	predParsed.trainClause = trainParsed.trainClause
+	a.NoError(createPredictionTable(predParsed, testDB, nil))
+}
