+++ pkg/sql/ir_generator_test.go
@@ -178,12 +178,12 @@ USING sqlflow_models.mymodel;`
 	modelDir, e := ioutil.TempDir("/tmp", "sqlflow_models")
 	a.Nil(e)
 	defer os.RemoveAll(modelDir)
-	stream := runExtendedSQL(`SELECT * FROM iris.train
+	stream := parseAndRunSQL(`SELECT * FROM iris.train
 TO TRAIN DNNClassifier
 WITH model.n_classes=3, model.hidden_units=[10,20]
 COLUMN sepal_length, sepal_width, petal_length, petal_width
 LABEL class
-INTO sqlflow_models.mymodel;`, testDB, modelDir, nil)
+INTO sqlflow_models.mymodel;`, modelDir, testDB)
 	a.True(goodStream(stream.ReadAll()))
 
 	// Test generate PredicrIR
@@ -211,19 +211,18 @@ func TestGenerateAnalyzeIR(t *testing.T) {
 	modelDir, e := ioutil.TempDir("/tmp", "sqlflow_models")
 	a.Nil(e)
 	defer os.RemoveAll(modelDir)
-	stream := runExtendedSQL(`
-	SELECT *
-	FROM iris.train
-	TO TRAIN xgboost.gbtree
-	WITH
-	    objective="multi:softprob",
-	    train.num_boost_round = 30,
-	    eta = 0.4,
-	    num_class = 3
-	COLUMN sepal_length, sepal_width, petal_length, petal_width
-	LABEL class
-	INTO sqlflow_models.my_xgboost_model;
-	`, testDB, modelDir, nil)
+	stream := parseAndRunSQL(`SELECT * FROM iris.train
+TO TRAIN xgboost.gbtree
+WITH
+	objective="multi:softprob",
+	train.num_boost_round = 30,
+	eta = 0.4,
+	num_class = 3
+COLUMN sepal_length, sepal_width, petal_length, petal_width
+LABEL class
+INTO sqlflow_models.my_xgboost_model;
+`, modelDir, testDB)
+	a.NoError(e)
 	a.True(goodStream(stream.ReadAll()))
 
 	// Test generate AnalyzeIR
